+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 This is a broad overview of the changes that have been made over the lifespan of this library.
 
-## 0.10.13 - 2022-05-04
+## v0.10.13 - 2022-05-05
 
 - Added py.typed file
+- Added some examples in a new examples directory
+    - This includes a benchmark with thefuzz
 
 ## v0.10.12 - 2022-05-04
 
@@ -16,7 +18,7 @@ This is a broad overview of the changes that have been made over the lifespan of
 - Adjusted sorting algorithm for functions that return multiple results
     - Should yield a bit better results when ratios are tied
 
-## 0.10.10 - 2022-05-03
+## v0.10.10 - 2022-05-03
 
 - Adjusted partial matching a tiny bit more
 - Adjusted tests accordingly